Fix some instances of -Wsign-compare
Fix remaining instances (other than chrome/browser/ash/) as a precursor
to reenable those warnings in //third_party/shell-encryption.

Bug: 1095122
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3921580
Commit-Queue: Henrique Ferreiro <hferreiro@igalia.com>
Cr-Commit-Position: refs/heads/main@{#1052332}

+++ chrome/browser/apps/app_service/publishers/arc_apps_unittest.cc
@@ -177,7 +177,7 @@ class ArcAppsPublisherTest : public testing::Test {
         });
 
     EXPECT_EQ(source.size(), target.size());
-    for (int i = 0; i < static_cast<int>(source.size()); i++) {
+    for (size_t i = 0; i < source.size(); i++) {
       EXPECT_EQ(*source[i], *target[i]);
     }
   }
@@ -417,10 +417,10 @@ TEST_F(ArcAppsPublisherTest,
 
   ASSERT_EQ(apps::State::SUCCESS, result.value_or(apps::State::FAILED));
 
-  ASSERT_EQ(file_system_instance()->handledUrlRequests().size(), 1);
+  ASSERT_EQ(file_system_instance()->handledUrlRequests().size(), 1u);
   auto& url_request = file_system_instance()->handledUrlRequests()[0];
   ASSERT_EQ(url_request->action_type, arc::mojom::ActionType::EDIT);
-  ASSERT_EQ(url_request->urls.size(), 1);
+  ASSERT_EQ(url_request->urls.size(), 1u);
   ASSERT_EQ(url_request->urls[0]->mime_type, "text/plain");
   ASSERT_TRUE(
       base::EndsWith(url_request->urls[0]->content_url.spec(), "test.txt"));
@@ -488,10 +488,10 @@ TEST_F(
 
   ASSERT_EQ(apps::State::SUCCESS, result.value_or(apps::State::FAILED));
 
-  ASSERT_EQ(file_system_instance()->handledUrlRequests().size(), 1);
+  ASSERT_EQ(file_system_instance()->handledUrlRequests().size(), 1u);
   auto& url_request = file_system_instance()->handledUrlRequests()[0];
   ASSERT_EQ(url_request->action_type, arc::mojom::ActionType::VIEW);
-  ASSERT_EQ(url_request->urls.size(), 2);
+  ASSERT_EQ(url_request->urls.size(), 2u);
   ASSERT_EQ(url_request->urls[0]->mime_type, "image/png");
   ASSERT_EQ(url_request->urls[1]->mime_type, "image/jpeg");
   ASSERT_TRUE(
@@ -527,10 +527,10 @@ TEST_F(ArcAppsPublisherTest,
 
   ASSERT_EQ(apps::State::SUCCESS, result.value_or(apps::State::FAILED));
 
-  ASSERT_EQ(file_system_instance()->handledUrlRequests().size(), 1);
+  ASSERT_EQ(file_system_instance()->handledUrlRequests().size(), 1u);
   auto& url_request = file_system_instance()->handledUrlRequests()[0];
   ASSERT_EQ(url_request->action_type, arc::mojom::ActionType::SEND);
-  ASSERT_EQ(url_request->urls.size(), 1);
+  ASSERT_EQ(url_request->urls.size(), 1u);
   ASSERT_EQ(url_request->urls[0]->mime_type, mime_type);
   ASSERT_TRUE(
       base::EndsWith(url_request->urls[0]->content_url.spec(), file_name));
@@ -559,10 +559,10 @@ TEST_F(ArcAppsPublisherTest, LaunchAppWithIntent_ShareFilesIntent_SendsExtras) {
       app_id, 0, std::move(intent), apps::LaunchSource::kFromFileManager,
       /*window_info=*/nullptr, base::DoNothing());
 
-  ASSERT_EQ(file_system_instance()->handledUrlRequests().size(), 1);
+  ASSERT_EQ(file_system_instance()->handledUrlRequests().size(), 1u);
   auto& url_request = file_system_instance()->handledUrlRequests()[0];
   ASSERT_EQ(url_request->action_type, arc::mojom::ActionType::SEND);
-  ASSERT_EQ(url_request->urls.size(), 1);
+  ASSERT_EQ(url_request->urls.size(), 1u);
   ASSERT_EQ(url_request->extras.value()[kTestExtraKey], kTestExtraValue);
   ASSERT_EQ(url_request->extras.value()["android.intent.extra.TEXT"],
             kTestIntentText);
