+++ chrome/browser/metrics/chromeos_metrics_provider_unittest.cc
@@ -5,6 +5,8 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 
 #include "chrome/browser/metrics/chromeos_metrics_provider.h"
 
+#include <stdint.h>
+
 #include <string>
 
 #include "ash/services/multidevice_setup/public/cpp/fake_multidevice_setup_client.h"
@@ -38,7 +40,7 @@ using Hardware = metrics::SystemProfileProto::Hardware;
 
 namespace {
 
-constexpr int kTpmFirmwareVersion = 100;
+constexpr uint64_t kTpmFirmwareVersion = 100;
 
 class FakeMultiDeviceSetupClientImplFactory
     : public ash::multidevice_setup::MultiDeviceSetupClientImpl::Factory {
